Rename base single and multi doc models
